Added accept reject functionality, profile storing functionality and offline functionality

+++ app/src/main/java/com/aagamshah/matchmate/data/remote/ProfileResponse.kt
@@ -29,7 +29,7 @@ data class Result (
 
 data class Dob (
     val date: String,
-    val age: Long
+    val age: Int
 )
 
 enum class Gender {
